Add test to get product

+++ application/mocks/application.go
@@ -91,10 +91,10 @@ func (mr *MockProductInterfaceMockRecorder) GetName() *gomock.Call {
 }
 
 // GetPrice mocks base method.
-func (m *MockProductInterface) GetPrice() float32 {
+func (m *MockProductInterface) GetPrice() float64 {
 	m.ctrl.T.Helper()
 	ret := m.ctrl.Call(m, "GetPrice")
-	ret0, _ := ret[0].(float32)
+	ret0, _ := ret[0].(float64)
 	return ret0
 }
 
@@ -157,7 +157,7 @@ func (m *MockProductServiceInterface) EXPECT() *MockProductServiceInterfaceMockR
 }
 
 // Create mocks base method.
-func (m *MockProductServiceInterface) Create(name string, price float32) (application.ProductInterface, error) {
+func (m *MockProductServiceInterface) Create(name string, price float64) (application.ProductInterface, error) {
 	m.ctrl.T.Helper()
 	ret := m.ctrl.Call(m, "Create", name, price)
 	ret0, _ := ret[0].(application.ProductInterface)
